docker deploy/django/mockups/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.0
Django==5.0.3
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
packaging==24.0
pillow==10.2.0
pytz==2024.1
requests==2.31.0
six==1.16.0
slugify==0.0.1
sqlparse==0.4.4
urllib3==2.2.1
